requirements.txt
pandas
requests
python-dotenv
nicegui